+++ tests/fixtures/cpp_issue113/remote.cpp
@@ -1,0 +1,3 @@
+#include "remote.h"
+using namespace rem;
+void Remote::Call() {}